Fixed EncryptionStream with large write sizes
When writing 64KB or more to an EncryptionStream before flushing, the
stream needs to break the data up into multiple boxes since
CryptoBox's max message size is 65535. It wasn't doing this, so
EncryptoBox was throwing an exception.

Also discovered that the `Compact` protocol's max msg size was
actually only (65535-16) because the size of the MAC was included in
the 2-byte size header. So I changed the header to _not_ include the
MAC size, making it simply the plaintext size.

Added a unit test to exercise sending large data and expose the bugs.