Fixed: Server stops accepting connections
The server didn't set up for accepting another connection until the
current connection finished the SecretHandshake. So if a connection
opened but didn't send any data, the server wouldn't accept any more.

While I was at it, I added a 15sec timeout waiting for a new
connection to finish the handshake.